requirement.txt
fastapi
uvicorn
pydantic
python-dotenv
openai
instructor